refactor: SelectJdbcTemplate, JdbcTemplate 통합
